chore(driver-interface): rename vehiclePlans
